The setlang view has some usability itches (wontfix)
Bij default, it redirects back to the same page.
So you don’t see anything has changed (even the language selector doesn’t move).

We could solve this by instead of using this form,
inserting links to the other language versions of this page.

But the form is actually useful because it sets the session variable.

So I now made it redirect to the home page by default,
after which the homepage redirects to the homepage in the newly chosen language.